requirements.txt
fastapi
uvicorn
requests
python-dotenv
beautifulsoup4
pypdf
sentence-transformers
faiss-cpu